mead/add_new_codes generate new migration file
