+++ rest-api/alembic/versions/9b93d00a35cc_add_new_biobank_codes.py
@@ -1,0 +1,72 @@
+"""add_new_biobank_codes
+
+Revision ID: 9b93d00a35cc
+Revises: 574efa4de1ba
+Create Date: 2018-01-26 07:59:58.808302
+
+"""
+from alembic import op
+import sqlalchemy as sa
+import model.utils
+
+
+from participant_enums import PhysicalMeasurementsStatus, QuestionnaireStatus, OrderStatus
+from participant_enums import WithdrawalStatus, SuspensionStatus
+from participant_enums import EnrollmentStatus, Race, SampleStatus, OrganizationType
+from participant_enums import MetricSetType, MetricsKey
+from model.site_enums import SiteStatus
+from model.code import CodeType
+
+# revision identifiers, used by Alembic.
+revision = '9b93d00a35cc'
+down_revision = '574efa4de1ba'
+branch_labels = None
+depends_on = None
+
+
+def upgrade(engine_name):
+    globals()["upgrade_%s" % engine_name]()
+
+
+def downgrade(engine_name):
+    globals()["downgrade_%s" % engine_name]()
+
+
+
+def upgrade_rdr():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('participant_summary', sa.Column('sample_order_status_1ps08', model.utils.Enum(OrderStatus), nullable=True))
+    op.add_column('participant_summary', sa.Column('sample_order_status_1ps08_time', model.utils.UTCDateTime(), nullable=True))
+    op.add_column('participant_summary', sa.Column('sample_order_status_1ss08', model.utils.Enum(OrderStatus), nullable=True))
+    op.add_column('participant_summary', sa.Column('sample_order_status_1ss08_time', model.utils.UTCDateTime(), nullable=True))
+    op.add_column('participant_summary', sa.Column('sample_status_1ps08', model.utils.Enum(SampleStatus), nullable=True))
+    op.add_column('participant_summary', sa.Column('sample_status_1ps08_time', model.utils.UTCDateTime(), nullable=True))
+    op.add_column('participant_summary', sa.Column('sample_status_1ss08', model.utils.Enum(SampleStatus), nullable=True))
+    op.add_column('participant_summary', sa.Column('sample_status_1ss08_time', model.utils.UTCDateTime(), nullable=True))
+    # ### end Alembic commands ###
+
+
+def downgrade_rdr():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_column('participant_summary', 'sample_status_1ss08_time')
+    op.drop_column('participant_summary', 'sample_status_1ss08')
+    op.drop_column('participant_summary', 'sample_status_1ps08_time')
+    op.drop_column('participant_summary', 'sample_status_1ps08')
+    op.drop_column('participant_summary', 'sample_order_status_1ss08_time')
+    op.drop_column('participant_summary', 'sample_order_status_1ss08')
+    op.drop_column('participant_summary', 'sample_order_status_1ps08_time')
+    op.drop_column('participant_summary', 'sample_order_status_1ps08')
+    # ### end Alembic commands ###
+
+
+def upgrade_metrics():
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
+
+def downgrade_metrics():
+    # ### commands auto generated by Alembic - please adjust! ###
+    pass
+    # ### end Alembic commands ###
+
