--- rest-api/alembic/versions/ec9339c5ef20_add_biobank_codes.py
@@ -1,57 +1,0 @@
-"""add_biobank_codes
-
-Revision ID: ec9339c5ef20
-Revises: 574efa4de1ba
-Create Date: 2018-01-25 16:29:52.206674
-
-"""
-from alembic import op
-import sqlalchemy as sa
-import model.utils
-
-
-from participant_enums import PhysicalMeasurementsStatus, QuestionnaireStatus, OrderStatus
-from participant_enums import WithdrawalStatus, SuspensionStatus
-from participant_enums import EnrollmentStatus, Race, SampleStatus, OrganizationType
-from participant_enums import MetricSetType, MetricsKey
-from model.site_enums import SiteStatus
-from model.code import CodeType
-
-# revision identifiers, used by Alembic.
-revision = 'ec9339c5ef20'
-down_revision = '574efa4de1ba'
-branch_labels = None
-depends_on = None
-
-def upgrade(engine_name):
-    globals()["upgrade_%s" % engine_name]()
-
-
-def downgrade(engine_name):
-    globals()["downgrade_%s" % engine_name]()
-
-
-
-def upgrade_rdr():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade_rdr():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def upgrade_metrics():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
-
-def downgrade_metrics():
-    # ### commands auto generated by Alembic - please adjust! ###
-    pass
-    # ### end Alembic commands ###
-
